pyproject.toml
[tool.poetry]
name = "weather-checker-application"
version = "0.1.0"
description = ""
authors = ["gilihagbi <hagbi.gili@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "0.21.0"
pandas = "^2.2.2"
requests = "^2.32.3"
streamlit = "0.89.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
